refactor: move function imageLoaded to utils
