+++ source/js/modules/full-page-scroll.js
@@ -1,15 +1,5 @@
 import throttle from "lodash/throttle";
-
-function imageLoaded(src, alt = ``) {
-  return new Promise((resolve) => {
-    const image = document.createElement(`img`);
-
-    image.setAttribute(`alt`, alt);
-    image.setAttribute(`src`, src);
-
-    image.addEventListener(`load`, () => resolve(image));
-  });
-}
+import {imageLoaded} from "../utils/resources";
 
 export default class FullPageScroll {
   constructor() {
